% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{boot_survey_error}
\alias{boot_survey_error}
\title{Function to estimate boot-strapped errors for annual index}
\usage{
boot_survey_error(best_model, boot_reps = 500, year, variables, pred_pa = 1,
  ob_CPUE, distribution = "normal")
}
\arguments{
\item{year}{Set of habitat variables to use in the modeling}

\item{variables}{Set of habitat variables to use in the modeling}

\item{pred_pa}{Predicted presence or absence at the transect}

\item{ob_CPUE}{Observed CPUE at the transect}

\item{distribution}{Either "normal" (the default) or "gamma"}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}

\item{par}{values for each of the parameters}
}
\description{
This function takes the best fitting model and resamples the data refitting and recomputing
the indices to come up with a bootstrapped error estimate for the annual index value.
}
\examples{
dist_xy()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
