% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{cdf_bound}
\alias{cdf_bound}
\title{This function calculates cdf of CPUE for predicting presence or absence in a habitat model}
\usage{
cdf_bound(variable, CPUE)
}
\arguments{
\item{variable}{Set of habitat variables used to predict presence or absence}

\item{CPUE}{Observed CPUE at the transect}
}
\description{
This function calculates the cumulative distribution of the CPUE of a species weighted across a variable.
It computes the upper 95% and lower 5% boundaries for the distribution of a species over a variable.
}
\examples{
cdf_bound()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
