% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{dist_xy}
\alias{dist_xy}
\title{This function estimates distance of one set of points to another}
\usage{
dist_xy(lat1, long1, lat2, long2, unit)
}
\arguments{
\item{lat1}{Latitude of point 1}

\item{long1}{Longitude of point 1}

\item{lat2}{Latitude of point 2}

\item{long2}{Longitude of point 2}

\item{unit}{either "km" or "m"}
}
\description{
This function takes two lat-long pairs and calculates the distance between
them in m or km
}
\examples{
dist_xy()
}
\keyword{distance}
