% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{hab_equation}
\alias{hab_equation}
\title{A function to compile the right side of a habitat equation}
\usage{
hab_equation(variables, form, par)
}
\arguments{
\item{variables}{Set of habitat variables to use in the modeling}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}

\item{par}{starting values for each of the parameters}
}
\description{
This function compiles text of equation for model. It takes the habitat variables
and makes them into a text version of the right side of an equation that is then read into the
habitat model function.
}
\examples{
hab_equation()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
