% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{HabModel}
\alias{HabModel}
\title{A function to predict CPUE from a habitat equation}
\usage{
HabModel(variables, form, par, pred_pa = 1, year)
}
\arguments{
\item{variables}{Set of habitat variables to use in the modeling}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}

\item{par}{values for each of the parameters}

\item{pred_pa}{the predicted presence or absence at each transect}

\item{year}{year of the data collection}
}
\description{
This function uses the habitat model equation and parameter values from
the hab_equation function to calculate a predicted CPUE. A dummy variable
for a year effect is added from the yearp function using the year of the 
survey for each of the stations. The function also #' can incorporate a 
presence-absence prediction (defaults to 1 for present) #' from an external 
step (usually from the presence_absence function)
}
\examples{
HabModel()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
