% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{index_calc}
\alias{index_calc}
\title{Function to estimate annual index}
\usage{
index_calc(parameter_ests, best_model, best_parameters, variables, year, minob,
  boot_reps)
}
\arguments{
\item{parameter_ests}{Bootstrap estimates of parameters}

\item{best_model}{for each variable, a form for the model equation (1, 2 or 3)}

\item{best_parameters}{fitted values for each of the parameters}

\item{variables}{Set of habitat variables to use in the modeling}

\item{year}{years of the survey}

\item{minob}{1/2 of the minimum CPUE observation for backtransformation of the data}

\item{ob_CPUE}{Observed CPUE at the transect}

\item{pred_pa}{Predicted presence or absence at the transect}
}
\description{
This function takes the best fitting model and resamples the data refitting and recomputing
the indices to come up with a bootstrapped error estimate for the annual index value.
}
\examples{
index_calc()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
