% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{iteration}
\alias{iteration}
\title{A wrapper function to find the best fitting MEHRSI habitat model}
\usage{
iteration(variables, form, ob_CPUE, pred_pa = 1, par, year,
  distribution = "normal")
}
\arguments{
\item{variables}{Set of habitat variables to use in the modeling}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}

\item{ob_CPUE}{Observed CPUE at the transect}

\item{pred_pa}{Predicted presence or absence at the transect}

\item{par}{values for each of the parameters}

\item{year}{Set of habitat variables to use in the modeling}

\item{distribution}{Either "normal" (the default) or "gamma"}
}
\description{
This function takes the habitat variables, observed CPUE, the equation form, starting 
parameters, the predicted presence or absence, the year, and the distribution. The
function fits an intital full model and then iteratively reduces the number of
parameters for each habitat variable. It then chooses the best-fitting reduced model
and repeats the process until there is no improvement in AIC with the removal of 
additional parameters or variables. The function then returns a list with the fitting
history and the best model results.
}
\examples{
iteration()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
