% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{modlike}
\alias{modlike}
\title{This function calculates negative log-liklihood for a habitat model}
\usage{
modlike(par, form, variables, ob_CPUE, pred_pa = 1, year,
  distribution = "normal")
}
\arguments{
\item{par}{values for each of the parameters}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}

\item{variables}{Set of habitat variables to use in the modeling}

\item{ob_CPUE}{Observed CPUE at the transect}

\item{pred_pa}{Predicted presence or absence at the transect}

\item{year}{Set of habitat variables to use in the modeling}

\item{distribution}{Either "normal" (the default) or "gamma"}
}
\description{
This function takes observed and predicted CPUE and calculates a negative log-likelhood for the data
it uses either a log-normal or gamma distribution. The model form, parameters, habitat
variables, presence or absence and year are passed through this function to the HabModel function which computes the
predicted values. It accepts "normal" or "gamma" as arguments for distribution.
}
\examples{
modlike()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
