% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{presence_absence}
\alias{presence_absence}
\title{This function estimates presence or absence at each transect}
\usage{
presence_absence(variables, limits)
}
\arguments{
\item{variables}{Variables used for determining presence or absence}

\item{limits}{Limits on the distribution of a species from cdf_limits function}
}
\description{
This function uses the cdf_limits function to calculate the limits of presence or absence 
for a species based on a variable of interest. For example, if the cdf_limits found that 
a species was only present at depths between 100 m and 200 m, this function would input 
the depths from a trawl survey haul and determine presence or absence or absence of the 
species at that site.
}
\examples{
presence_absence()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
