% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{spatial_resids}
\alias{spatial_resids}
\title{Function to spatial autocorrelation in residuals}
\usage{
spatial_resids(longitude, latitude, residuals, pred_CPUE, ob_CPUE,
  region = "GOA")
}
\arguments{
\item{longitude}{Longitude of haul}

\item{latitude}{latitude of haul}

\item{residuals}{residuals of the best fitting model}

\item{pred_CPUE}{predicted CPUE of the best fitting model}

\item{ob_CPUE}{Observed CPUE at the transect}

\item{region}{either "GOA" (default) or "AI"}
}
\description{
This function takes the best fitting model and resamples the data refitting and recomputing
the indices to come up with a bootstrapped error estimate for the annual index value.
}
\examples{
spatial_resids()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
