% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHRSI.R
\name{yearp}
\alias{yearp}
\title{A function to create a dummy year variable for a habitat equation}
\usage{
yearp(year, par, form)
}
\arguments{
\item{year}{Set of habitat variables to use in the modeling}

\item{par}{values for each of the parameters}

\item{form}{for each variable, a form for the model equation (1, 2 or 3)}
}
\description{
This function creates a dummy variable for a year effect that can be
added to the habitat model equation.
}
\examples{
yearp()
}
\keyword{abundance}
\keyword{habitat}
\keyword{index}
\keyword{model,}
\keyword{survey}
